major code erorr and eslint issues fixed
